requirements.txt
ariadne==0.21
bcrypt
Flask-Cors==4.0.0
Flask-RESTful==0.3.10
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
flask-talisman==1.1.0
Flask==3.0.0
greenlet==3.0.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
psycopg2-binary==2.9.9
psycopg2==2.9.9
pyjwt
python-dotenv==1.0.0
python-socketio==5.11.0
simple-websocket==1.0.0
SQLAlchemy==2.0.25
Werkzeug==3.0.0
marshmallow==3.20.2
Flask-Mail==0.9.1
requests==2.31.0
selenium==4.21.0
webdriver-manager==4.0.1
Babel==2.15.0
nutree==0.8.0
python-dateutil==2.9.0.post0
waitress==3.0.0
click==8.1.7
# pytest==8.2.2
firebase-admin==6.5.0
Flask-APScheduler==1.13.1
qrcode==7.4.2
deepmerge==2.0
# 
twilio==9.3.1
